pyproject.toml
[tool.poetry]
name = "milerobo"
version = "0.1.0"
description = ""
authors = ["houfu <houfu@outlook.stg>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
python-dotenv = "^0.21.1"
openai = "^0.26.5"
langchain = "^0.0.84"
weaviate-client = "^3.11.0"


[tool.poetry.group.scraper.dependencies]
scrapy = "^2.8.0"
beautifulsoup4 = "^4.11.2"
html5lib = "^1.1"
weaviate-client = "^3.11.0"
langchain = "^0.0.84"
spacy = "^3.5.0"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl"}
tiktoken = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
